Rename files to use UpperCamelCaseCluster/Logic for general-diagnostics-server
