+++ src/app/clusters/general-diagnostics-server/app_config_dependent_sources.cmake
@@ -24,6 +24,6 @@ TARGET_SOURCES(
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
-    "${CLUSTER_DIR}/general-diagnostics-cluster.h"
-    "${CLUSTER_DIR}/general-diagnostics-cluster.cpp"
+    "${CLUSTER_DIR}/GeneralDiagnosticsCluster.h"
+    "${CLUSTER_DIR}/GeneralDiagnosticsCluster.cpp"
 )
